ai-service/requirements.txt
flask
flask_cors
pandas
scikit-learn
pymysql
python-dotenv
google-genai
